# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample Unicef server."
  version: "1.0.0"
  title: "Unicef"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: ""
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /api/processorder:
    post:
      tags:
      - "Orders"
      summary: "Process order for point calculation"
      description: ""
      operationId: "processOrder"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          properties:
            orderId:
              type: "string"
              example: "605059f391ad10274d40f663"
            redeemedPoints:
              type: "number"
              example: 50
            countryCode:
              type: "string"
              example: "IND"
            pointSource:
              type: "string"
              example: "Online"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: true
              message:
                type: "string"
                example: "Order processed successfully!!"
              discountAmount:
                type: "integer"
                example: 50
              finalPrice:
                type: "integer"
                example: 500
              totalEarnedPoints:
                type: "integer"
                example: 5
        "500":
          description: "Invalid input"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "Error in order processing!!"
              discountAmount:
                type: "integer"
                example: 0
              finalPrice:
                type: "integer"
                example: 0
              totalEarnedPoints:
                type: "integer"
                example: 0
        "400":
          description: "parameter is missing"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "countryCode parameter is missing"
  /api/pointsupdate:
    post:
      tags:
      - "Orders"
      summary: "Points update "
      description: ""
      operationId: "pointsUpdate"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          properties:
            orderId:
              type: "string"
              example: "605059f391ad10274d40f663"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: true
              message:
                type: "string"
                example: "Point updated successfully!!"
             
        "500":
          description: "Invalid input"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "Error in point updation processing!!"
             
        "400":
          description: "parameter is missing"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "orderId parameter is missing"
...
  /api/redeemdetails:
    post:
      tags:
      - "Points"
      summary: "Points redeemtion "
      description: ""
      operationId: "pointsRedeemtion"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          properties:
            residentId:
              type: "string"
              example: "605059f391ad10274d40f663"
            redeemedPoints:
              type: "number"
              example: 50
      responses:
         "200":
           description: "successful operation"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: true
               message:
                 type: "string"
                 example: "User has available redeempoints"
         "500":
           description: "Invalid input"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "Sorry, You don't have availablepoints to redeem"
         "400":
           description: "parameter is missing"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "residentId parameter is missing"
  /api/user/transactiondetails/{residentId}:
    get:
      tags:
      - "Points"
      summary: "Points transaction details "
      description: ""
      operationId: "pointsTransaction"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "residentId"
        in: "path"
        description: "Resident id required"
        required: true
      responses:
         "200":
           description: "successful operation"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: true
               message:
                 type: "string"
                 example: "User transaction details"
               data:
                 type: "array"
                 items:
                   $ref: "#/definitions/PointsAudit"
         "500":
           description: "Invalid input"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "Sorry, You don't have transaction details"
               data:
                 type: "array"
                 items:
                   type: object
                   example: null
         "400":
           description: "parameter is missing"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "residentId parameter is missing"
  /api/user/points/{residentId}:
    get:
      tags:
      - "Points"
      summary: "Points details "
      description: ""
      operationId: "pointsDetails"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "residentId"
        in: "path"
        description: "Resident id required"
        required: true
      responses:
         "200":
           description: "successful operation"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: true
               message:
                 type: "string"
                 example: "User points details available"
               data:
                 type: "object"
                 example: { "totalLapsedPoints":"100","totalAvailablePoints":"500","totalEarnedPoints":"50","totalRedeemedPoints":"20"}
         "500":
           description: "Invalid input"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "Sorry,User has no point"
               data:
                 type: "object"
                 example: null
                 
         "400":
           description: "parameter is missing"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "residentId parameter is missing"
...
definitions:
...
  PointsAudit:
    type: "object"
    properties:
      residentId:
        type: "string"
        example: "6048708c946251d02c88f329"
      orderId:
        type: "string"
        example: "6048708c946251d02c88f329"
      redeemedPoints:
        type: "integer"
        format: "int32"
        example: 100
      earnedPoints:
        type: "integer"
        format: "int32"
        example: 25
      availablePoints:
        type: "integer"
        format: "int32"
        example: 125
      isActive:
        type: "boolean"
        default: false
      isLapsed:
        type: "boolean"
        default: false
      pointSource:
        type: "string"
        example: "online"
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
    xml:
      name: "PointsAudit"
...